Move config.ts to formatter/ dir
